requirements.txt
black~=24.4.2
click~=8.1.7
isort~=5.13.2
mypy==1.10.1
pydocstyle[toml]==6.3.0
pylint==3.2.4
pytest~=8.2.2
pytest-cov~=5.0.0
setuptools_scm[toml]>=8.1.0
twine~=5.1.0
xenon==0.9.1

# Dependencies
e3nn
gatr 
geomloss
hydra-core
h5py
lab_gatr @ git+https://github.com/sukjulian/lab-gatr.git
potpourri3d
prettytable
pykeops
pyvista
torch==2.5.1  # for PyG
torch_geometric
torch-scatter
torch-cluster
--find-links https://data.pyg.org/whl/torch-2.5.1+cu124.html  # update upon new release(s)
trimesh
wandb
xformers


# Static code analysis
autoflake
pre-commit
types-tqdm


# For hyperparameter tuning
optuna

dotenv